# Conflicts:
#	csharp/SEP3UI/Pages/OrderList.razor

+++ java/src/main/java/database/daomodel/BookDAO.java
@@ -2,12 +2,13 @@ package database.daomodel;
 
 import model.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 
 public interface BookDAO {
-    Book create(String name, String description, double price, Category category, int quantity,
+    Book create(String name, String description, BigDecimal price, Category category, int quantity,
                 String imgFilePath, String ISBN, List<Author> authors, String language,
                 List<Genre> genre, LocalDate publicationDate);
     Book read(String ISBN, int id);
